requirements.txt
streamlit>=1.33.0
streamlit-extras>=0.4.0
requests>=2.32.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
pymongo>=4.6.1
SQLAlchemy>=2.0.29
PyMySQL>=1.1.0
openai>=1.23.0
python-dotenv>=1.0.1
